fix: the excluded finding reads as excluded in every renderer
Codex review of the authoritative pipeline found two more consumers still
reading the unfiltered change set, both reproduced before fixing.

**P1 — SARIF, JUnit and the review digest contradicted their own verdict.**
Each classifies every `result.changes` entry by its effective kind verdict,
so a proven-out-of-contract type-size change produced, on one run:

    verdict: NO_CHANGE   SARIF level: error   JUnit failures: 1
    digest: "safe to merge"  ... "Top impacted symbols: Internal"

SARIF now annotates such a finding `level: note` and states the relevance
and reason code that explain it; JUnit gives it a passing `<testcase>` the
same way `_is_failure` already handles a scoped-out finding; the digest's
impacted list is over what actually gated. Downgraded, never dropped --
ADR-049 D9 requires the fact to land in exactly one *visible* outcome.

**P2 — `scan --against` rows omitted the canonical decision pair.** They
carried relevance/reason/assurance but not
`compatibility_evaluation_status`/`compatibility_decision`, so a scan row
and the `compare` finding for the same fact could not be compared field by
field -- the divergence ADR-049 section 6.4 exists to prevent. Added, with
`null` as the required value for an unscored row, and the parity gate's own
shared-field list extended so it checks them from now on. Deliberately not
`gate_contribution`: that is a property of a severity gate this command does
not run.

Also splits `contract_evaluation.py` (1972 lines, thirty short of the hard
cap) at the seam CodeRabbit identified. `contract_scoped_promotion.py` (466
lines) takes ADR-049 section 4.3's explicit-scope tier and everything that
tier implies -- the promotion, the finding's own decision, the monotonic
verdict recompute, the missing-contract contribution, the receipt refresh.
The seam is clean in both directions: the tail needed two constants from the
head and the head needed nothing back, so consumers now import from the
module that owns each function rather than through a re-export. Parent left
at 1552 lines; no new import cycle.
